api: initial implementation of headless API (Bug 1941363)
Build out the basic functionality of the headless API for Lando.
Using django-ninja we define two API endpoints, to POST automation
jobs and GET job statuses after submission. The API endpoints take
a set of `actions` defined in the request body which are stored in
the database for processing by a worker. Authentication is handled by
an API key associated with a user profile. A single action, `add-commit`
is implemented which can be used to test adding patches to the repo
as commits.